pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "voicebot"
version = "0.1.0"
description = "German speech transcription and keyword detection voicebot with Flask backend and Streamlit frontend."
readme = "README.md"
requires-python = "==3.12.10"
authors = [
  { name = "Nagaraju Javisetty", email = "nagajavisetty143@gmail.com" }
]

dependencies = [
  # Backend dependencies
  "fastapi==0.122.0",
    "uvicorn== 0.38.0",
  "torch== 2.9.1",
  "transformers== 4.41.1",
  "pydub== 0.25.1",
    "python-multipart==0.0.20",
  # Frontend dependencies
  "streamlit==1.35.0",
  "requests==2.32.3",
  "streamlit-audiorecorder==0.0.6",

]

[project.optional-dependencies]
dev = [
  "pytest",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["frontend", "backend"]

[project.urls]
Homepage = "https://example.com/voicebot"
